feat: prelaunch changes
