+++ src/pages/index.tsx
@@ -1,34 +1,12 @@
-import { GetStaticProps, GetStaticPropsResult, NextPage } from "next";
+import { GetStaticProps, NextPage } from "next";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 
-import AppLayout from "@/components/SharedComponents/AppLayout";
-import TeachersTab from "@/components/GenericPagesComponents/TeachersTab";
-import HomePageTabImageNav from "@/components/GenericPagesComponents/HomePageTabImageNav";
-import { HomePageLowerView } from "@/components/GenericPagesViews/HomePageLower/HomePageLower.view";
-import getPageProps from "@/node-lib/getPageProps";
-import curriculumApi2023, {
-  TeachersHomePageData,
-} from "@/node-lib/curriculum-api-2023";
-import {
-  getBlogPosts,
-  SerializedPost,
-} from "@/pages-helpers/home/getBlogPosts";
-import { HomePage } from "@/common-lib/cms-types";
-import { DEFAULT_SEO_PROPS } from "@/browser-lib/seo/Seo";
+import { HomePageProps } from "@/pages-helpers/home/getBlogPosts";
+import AiPage, { getStaticProps as getStaticPropsAi } from "@/pages/ai/index";
 
-export type TeachersHomePageProps = HomePageProps & {
-  curriculumData: TeachersHomePageData;
-};
-
-export type HomePageProps = {
-  pageData: HomePage;
-  posts: SerializedPost[];
-};
-
-const Home: NextPage<TeachersHomePageProps> = (props) => {
+const Home: NextPage<HomePageProps> = (props) => {
   const router = useRouter();
-  const { curriculumData, posts } = props;
 
   useEffect(() => {
     // clientside redirect for old tabs implementation
@@ -39,49 +17,13 @@ const Home: NextPage<TeachersHomePageProps> = (props) => {
     } else if (window.location.href.includes("#ai")) {
       router.push("/ai");
     } else if (window.location.href.includes("#teachers")) {
-      router.push("/");
+      router.push("/teachers");
     }
   }, [router]);
 
-  return (
-    <AppLayout seoProps={DEFAULT_SEO_PROPS} $background={"white"}>
-      <HomePageTabImageNav current={"teachers"} />
-      <TeachersTab keyStages={curriculumData.keyStages} aria-current="page" />
-      <HomePageLowerView posts={posts} />
-    </AppLayout>
-  );
+  return <AiPage {...props} />;
 };
 
-export const getStaticProps: GetStaticProps<HomePageProps> = async (
-  context,
-) => {
-  return getPageProps({
-    page: "teachers-home-page::getStaticProps",
-    context,
-    getProps: async () => {
-      const { pageData, posts } = await getBlogPosts(
-        context.preview === true,
-        5,
-      );
-
-      if (!pageData) {
-        return {
-          notFound: true,
-        };
-      }
-
-      const curriculumData = await curriculumApi2023.teachersHomePage();
-
-      const results: GetStaticPropsResult<TeachersHomePageProps> = {
-        props: {
-          pageData,
-          curriculumData,
-          posts,
-        },
-      };
-      return results;
-    },
-  });
-};
+export const getStaticProps: GetStaticProps<HomePageProps> = getStaticPropsAi;
 
 export default Home;
